+++ simple-interest/calculator/simple_interest_calculator.go
